Extract constraint Args to schemas
Generated-by: Claude Opus 4.6
