+++ src/guidellm/scheduler/constraints/request.py
@@ -13,10 +13,6 @@ from typing import Any, Literal, cast
 
 from pydantic import Field
 
-from guidellm.scheduler.constraints.args import (
-    ConstraintArgs,
-    PositiveNumOrList,
-)
 from guidellm.scheduler.constraints.constraint import (
     Constraint,
     PydanticConstraintInitializer,
@@ -28,6 +24,10 @@ from guidellm.scheduler.schemas import (
     SchedulerUpdateAction,
 )
 from guidellm.schemas import RequestInfo, StandardBaseModel
+from guidellm.schemas.scheduler.constraints import (
+    MaxDurationConstraintArgs,
+    MaxRequestsConstraintArgs,
+)
 from guidellm.utils.mixins import InfoMixin
 
 __all__ = [
@@ -37,44 +37,6 @@ __all__ = [
     "MaxRequestsConstraintArgs",
     "RequestsExhaustedConstraint",
 ]
-
-
-@ConstraintArgs.register("max_duration")
-class MaxDurationConstraintArgs(ConstraintArgs):
-    """
-    Arguments for maximum duration constraint.
-
-    Limits benchmark execution time per strategy.
-
-    :cvar kind: Always "max_duration"
-    """
-
-    kind: Literal["max_duration"] = Field(
-        default="max_duration",
-        description="Constraint type discriminator",
-    )
-    seconds: PositiveNumOrList = Field(
-        description="Maximum duration in seconds before stopping execution",
-    )
-
-
-@ConstraintArgs.register("max_requests")
-class MaxRequestsConstraintArgs(ConstraintArgs):
-    """
-    Arguments for maximum request count constraint.
-
-    Limits the number of requests processed per strategy.
-
-    :cvar kind: Always "max_requests"
-    """
-
-    kind: Literal["max_requests"] = Field(
-        default="max_requests",
-        description="Constraint type discriminator",
-    )
-    count: PositiveNumOrList = Field(
-        description="Maximum number of requests before stopping execution",
-    )
 
 
 @ConstraintsInitializerFactory.register("max_requests")
